# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Caffe-Backend",
        "version": "0.1"
    },
    "paths": {
...
        "/api/produtos": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar produtos",
                "operationId": "cab8aec265fc5e5ab4a1d09399ead918",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página (opcional)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a lista de produtos"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Adicionar produto",
                "operationId": "4ebb3451d106364b9fd624ac3d4265ab",
                "requestBody": {
                    "description": "Informações do novo produto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nome_produto",
                                    "Preco_compra",
                                    "Preco_venda",
                                    "Descricao"
                                ],
                                "properties": {
                                    "Nome_produto": {
                                        "type": "string"
                                    },
                                    "Preco_compra": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "Preco_venda": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "Descricao": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto adicionado com sucesso"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/produtos/{id}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Mostrar detalhes do produto",
                "operationId": "160123aa3766f9eb53e8131219d24ae5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna os detalhes do produto"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualizar produto",
                "operationId": "41d748c0b2810d1c7d818f8928cf1441",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Informações do produto a serem atualizadas",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Nome_produto": {
                                        "type": "string"
                                    },
                                    "Preco_compra": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "Preco_venda": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "Descricao": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        }
    }
}